Load pots on all pages
